* commit '1d68ad6f2238a0b9aa7f8bcd1afc462f4d2f5350':
  develop price range filter
  Develop price range Screen
  filter done except price range
  apply filters

+++ app/src/main/java/com/hexagram/febys/models/api/filters/Range.kt
@@ -5,6 +5,6 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Range(
-    val max: Int,
-    val min: Int
+    val max: Double,
+    val min: Double
 ) : Parcelable
